+++ io_scene_habitatb/__init__.py
@@ -136,6 +136,9 @@ def menu_func_export_w(self, context):
 # POS
 def menu_func_import_pos(self, context):
     self.layout.operator(io_ops.ImportPOS.bl_idname, text="Re-Volt Position Nodes (.pan)")
+# INF
+def menu_func_import_fin(self, context):
+    self.layout.operator(io_ops.ImportFIN.bl_idname, text="Re-Volt Instances (.fin)")
 
 # CAR
 def menu_func_import_car(self, context):
@@ -148,7 +151,9 @@ def register():
     bpy.types.INFO_MT_file_import.append(menu_func_import_ncp)
     bpy.types.INFO_MT_file_import.append(menu_func_import_w)
     bpy.types.INFO_MT_file_import.append(menu_func_import_pos)
+    bpy.types.INFO_MT_file_import.append(menu_func_import_fin)
     bpy.types.INFO_MT_file_import.append(menu_func_import_car)
+    
     bpy.types.INFO_MT_file_export.append(menu_func_export_prm)
     bpy.types.INFO_MT_file_export.append(menu_func_export_ncp)
     bpy.types.INFO_MT_file_export.append(menu_func_export_w)
@@ -165,7 +170,9 @@ def unregister():
     bpy.types.INFO_MT_file_import.remove(menu_func_import_ncp)
     bpy.types.INFO_MT_file_import.remove(menu_func_import_w)
     bpy.types.INFO_MT_file_import.remove(menu_func_import_pos)
+    bpy.types.INFO_MT_file_import.remove(menu_func_import_fin)
     bpy.types.INFO_MT_file_import.remove(menu_func_import_car)
+
     bpy.types.INFO_MT_file_export.remove(menu_func_export_ncp)
     bpy.types.INFO_MT_file_export.remove(menu_func_export_w)
     bpy.types.INFO_MT_file_export.remove(menu_func_export_prm)